incorporated comments from pull request
